This commit introduces multi-dimensional arrays.
At runtime a multi-dimensional array is represented by a large one-dimensional array which guarantees good locality and non-jaggedness. In order to guarantee these two properties the array type, `[t]`, has been changed to `ArrayN<t>`, where `N` is the dimensionality of an array. Note that `Array2<t>` (a matrix) is not equal to `Array1<Array1<t>>` (an array of arrays), nor is either a subtype of the other.

Multi-dimensional arrays are indexed using a comma-separated list of integers, `a[i, j, k]`, and may not be partially indexed, except for when checking the size of a dimension.

Size checking of multi-dimensional arrays is performed similar to how it is done on arrays of arrays. `|a|` is the size of the array in the first dimension, `|a[i]|` is the size of the array in the second dimension and `|a[i, j]|` is the size of the array in the third dimension. Note that the actual values of `i`and `j` doesn't matter, only the length of the index list. This is likely the weakest part of this design and I'm open for suggestions.

When creating a multi-dimensional array the sizes of all dimensions must be provided. For example: `new Array2<int>(10, 20) ` creates a new array with 10 rows and 20 columns.

In a multi-dimensional array literal, like `[[1,2,3], [4,5,6]]`, all sub-arrays at the same level must be of the same type _and size_. Furthermore, expressions with array types are _not_ allowed inside array literals, unless they themselves are literals. This allows multi-dimensional array literals to be assigned an `ArrayN<t>` type instead of an array-of-arrays type.

This pull-request also fixes indexing of arrays-of-arrays so that it can be done like `a[i][j]` instead of `(a[i])[j]`.

Finally, the notation `[t]` is kept as an alias for `Array1<t>` in order to maintain backwards compatability. For convenience, `Array<t>` is also provided as an alias for the type `Array1<t>`.