fixes removal of breathe AST nodes
the breathe keyword has been removed from the language. however, there
is still some places that still use the breathe keyword. this is a
mistake and this commit removes those places, basically the AST node for
breathe and some utility functions that do no need the AST node anymore.
this commit fixes #487