Fix tracing function for ParTs (#461)
* fix tracing ParT types

the tracing function for the ParT types was not correctly generated and
caused runtime errors. The issues that have been fixed are shown below
and this closes #393:

- fix >> combinator with correct trace function
  the generated trace function was not correct, causing runtime crashes

- fix trace function for extract combinator was not correct and was
  tracing array_types. the trace function of the extract combinator was
  always set to be an array; this is not the case, as that trace
  function is used to indicate the trace function of the returned array.
  the function function gets the runtime base type of party, e.g. `Par
  t` has runtime type `t` and a `Par (Par t)` has runtime type `Par t`,
  since the array will contain `Par t`s.

- fix join combinator. removed assertion. In the following scenario
  (P p1 p2), an assertion tests that the tracing function of p1 and p2
  should be the same. this is not true, as p1 may be a (V 32) and p2 may
  be a (F f), which should have different tracing functions.

* refactor tests to not-used `extract`

refactor tests in ParT to avoid the use of the `extract` function. this
is due to an issue (#434) related to futures. the sequence test omits
the use of extract, and the liftf test has been re-written to avoid that
as well. tests that cannot avoid the use of `extract` have been disabled
and will be re-enable as soon as this works.

* remove `extract` from the language

* remove `extract` from documentation