Moved embed tests to own suite (#498)
* Moved embed tests to own suite

For some time, the test `embed-tl.enc` has been failing. The problem
seems to be a missing Makefile that builds the accompanying C code. This
commit fixes this omission and moves all embed tests to their own test
suite (so that the Makefile is together with the tests that is more
likely to need them).

* Added missing space

* Added check for empty CC and std=c11