Compiler support for literate programming
This commit adds support for [literate
programming](https://en.wikipedia.org/wiki/Literate_programming) in
Encore. This commit message is such a program, and can be compiled by
passing the `--literate` flag to the compiler. The compiler will then
filter out all the text that is not wrapped in code block delimiters
like so (this is org-mode syntax for literate programming):

 #+begin_src encore
 class Main
   def main() : void
     print "Literate programming is cool"
 #+end_src encore

Code blocks that don't specify encore will still be assumed to be part
of the encore code. Specifying a different language will make the code
block be ignored by the compiler:

 #+begin_src C
   void main(){
     puts "This will be ignored";
   }
 #+end_src C

This new feature is documented in `--help`, and there are tests in the
suite `encore/literal`. This could possibly be a way to write
documentation or tutorials that can be machine-checked for
well-formedness.