update context after future `get`
this commit updates the context in the method `encore_alloc` to solve
the problem of using an stale context. this commit also includes a test
case that checks that this issue won't happen
again (`refreshContext.enc`). as a final consequence, fixing this issue
solves the problem with the `extract` combinator which has now been
re-added. this commit fixes #434.