Backend support for intersection types
Fixes #348. This commit adds backend support for intersection types.
Intersection types take the form `T1 & T2` and should be interpreted as
a type that is **both** `T1` and `T2`. For now, this is only used as a
mechanism for handling the types of branching expressions (`if` and
`match`). For example, the `if` statement

```
if b then
  e1 : Foo
else
  e2 : Bar
```

will have the type `Foo & Bar`, meaning any operations performed on it
must be available in both `Foo` and `Bar` (in practice this is only true
if `Foo` and `Bar` has an overlap in their included capabilities). The
typechecker will try to get rid of intersection types when it can. For
example, if one of the branches has a type that is a subtype of all
other branches, the resulting type will be that type. If a class type is
encountered, its capability will be considered.

See `src/tests/encore/intersection` for tests and examples. As this is
not available to the programmer there is no documentation. When we find
a new home for the documentation (or if we decide to introduce
intersection types to the surface language) I can write something about
it.