Support for inference of type arguments
This commit adds support for inferring the type arguments of polymorphic
functions. Note that it currently only works for function calls, and not
for passing polymorphic functions around. The following example will
illustrate:

```
def bump(x : int) : int
  x + 1

def apply<a, b>(fn: a -> b, x : a) : b
  fn(x)

def id<a>(x : a) : a
  x

...

-- Explicit version (required before this commit):
apply<int, int>(bump, 42);

-- Inferred version (provided by this commit):
apply(bump, 42);

-- Limitation when using functions as values:
apply(id<int>, 42); -- (type argument of id must be provided)
```

All tests of polymorphic functions have been duplicated with inferring
versions. Failing tests have been added as well.

This commit also fixes a bug where the runtime type arguments for
parametric functions were always `ENCORE_PRIMITIVE` when a function was
used as a value. The cause was that the type arguments were resolved but
never saved in the AST-node, meaning that they appeared as
unresolved (which defaults to `ENCORE_PRIMITIVE`).