Improve parser behavior
This commit improves the backtracking behavior of the parser to give
error messages closer to where the error message actually is. Among
other minor improvements, the major contribution of this commit is
improvement in the following two cases:

* Syntax errors in an `else` branch will now be identified, instead of
  complaining that there is an "unexpected 'e'" in the `else` keyword.

* Syntax errors in the body of a `let x = 42 y = 13 in ...` will now be
  identified, instead of guessing that there should be a semi-colon
  between `42` and `y`.

A known remaining issue is that syntax errors in the arguments of
function calls will have error messages that point to the name of the
function rather than the actual error.

Fixes #542. Fixes #300. Should make a lot of people happy :)