Make sure embedded types have the same kind
This commit fixes a bug where checking if two embedded types had the
same kind would fail. For example, when assigning `x = y`, `x` and `y`
must be of the same kind. If `x` is an array of type `[t]`, `y` must be
an array of type `[t']` that where `t'` has the same kind as `t`. If `t`
was an embedded type, this would fail.

Fixes #575.