Let traits with closures be multiply included
This commit fixes the error where a trait defining a closure would cause
a linking error when included multiple times. This was because the
function representing the closure body was created once for each
including class (due to flattening). The solution was to make these
functions `static`, which also makes sense as they are method local.

Fixes #353.