Support for println
This commit adds support for `println` which works like calling `print`
as a function, but adds a newline to the end of the string. For
backwards compatibility reasons, the expression `print e` will still
print `e` with a newline (it is equivalent to `println("{}", e)`).
`println e` is malformed. Tests with explicit newlines have been updated
to use `println`

This commit also resolves the ambiguity where `print("foo")` would add a
newline, whereas `print("foo: {}", 42)` would not. Using `print` *as a
function* (i.e. with parenthesis around the arguments) will never add a
newline, whereas `println` will.

Fixes #581.