Support for mutual trait requirements
This commit adds support for letting a trait provide a method required
by another trait. For now, the typechecker is overly cautious and
disallows a trait to provide a method if it can access any mutable
fields that are not mutable in the requiring trait. The provider may
however access additional `val` fields. The rationale behind this is
that the "footprint" of a trait should be visible in the trait alone, so
that a trait that appears to only perform reads actually only performs
reads even if it requires a method from elsewhere. Later this might be
relaxed to only be mandatory for `read` capabilities.

See the tests `traitProvide*.enc` for examples of how mutual require
works and doesn't.

Fixes #448.