Relax constraints on mutual trait requirements
This commit relaxes the constraints on mutual trait requirements so that
a trait `P` *can* provide methods to a trait `R`, even when `P` requires
fields that are not in `R`. The remaining constraint is that `P` may not
see any fields as `var` fields that are required as `val` fields in `R`.
Tests have been updated.