Support for method overriding (Kappa ready)
This commit adds support for method overriding of concrete methods. The
same rules apply regardless of whether the overridden method is
abstract (`require`d) or concrete (`def`ed), meaning that some old
programs with overriding of abstract methods might need to be updated.

The basic idea is that a method `m` in a class `C` is typechecked once
with `this : C`, and then once per included trait `T` that has a method
that is overridden by `m`. Since `m` may access fields or methods that
are not in `T`, the requirements of `T` can be extended on inclusion:

```
trait T
  def m() : int
    42

passive class C : T(f, foo())
  f : int

  def foo() : int
    this.f

  def m() : int
    this.foo()
```

Note that the program would not typecheck for `passive class C : T`, as
`m` in `C` cannot be typechecked with `this : T` (since `foo` is not in
the interface of `T`). The type `T(f, foo())` should be interpreted as
an abstract trait which extends `T` with a field `f` and a (`require`d)
method `foo`. It can only appear where a trait is included in a trait.

Method overriding follows the same rules as for mutual method
requirements. A field that is marked as immutable (`val`) in the
included trait may not be mutated in an overriding method. This is
captured by the checking outlined above, and is compliant with the
future integration with Kappa.

See tests `traits/overriding*` and `traits/expr.enc` for examples.