Fix bug in test script (#622)
This commit fixes a bug where having an ignored file in a test suite
would prevent running that suite on its own. The important change is the
one in `src/tests/bin/test`. This also enables a test that was
previously erroneously disabled (and fixes a bug in that test).
`IGNORED_FILES.grep` has been changed to get rid of some warnings.