Fix inter-module method calling bug
This commit fixes a bug where calling methods and reading fields across
modules could result in a typechecking error. The cause of the bug was
that the types in a method or field did not get their namespaces
translated to match the module from which the lookup was done. A test
case for catching similar bugs has been updated to also cover these
cases. Fixes #626.