Support for handling unimported types
This commit adds support for calling functions or methods that return a
type that is not in scope. This type cannot be used to call methods or
access fields but can for example be passed back to the object from
which it was obtained or to another module that knows of this type. See
tests `HiddenType*` and `ModuleInTheMiddle` for examples.

Implementation-wise, a namespace can now be either explicit or implicit.
An explicit namespace is one that has a name in the current module (i.e.
`Foo.Bar.Baz`), whereas an implicit namespace is one that does not have
a name. When handling a value of a type that is not in scope, this type
will have an implicit namespace (printed as `<path/to/Source.enc>`) that
can be used to distinguish it from other types with the same name.