Fixed overriding bug regarding polymorphism
This commit fixes a bug where overridden methods from a polymorphic
trait were typechecked with `this` as the generic trait (e.g. `T<a>`)
rather than the concrete trait it was included as (e.g. `T<int>`),
causing problems when `this` was (correctly) used as the second type.
See `overridingPoly.enc` for a test case.