Don't crash when extending with missing method
This commit fixes a bug where extending an included trait with a method
that was not present in the including class crashed the compiler:

```
trait T
passive class C : T(foo()) {}
```

It would also crash even if `foo` was provided by another trait.

The fix is a bit involved since the source of the bug was a chicken and
egg problem: in order to typecheck the inclusion `passive class C :
T(foo()) + S`, one needs to resolve `T + S` and also be able to do
lookups on `T + S`, but in order to do lookups on `T + S`, the
capability has to be resolved first. The fix was to first resolve `T +
S` and then typecheck `T(foo()) + S` in a special environment where the
trait inclusion of `C` was the resolved version of `T + S` (without
extensions).

Fixes #643.