Fix bug using type parameters (#637)
This commit fixes a bug where including a polymorphic trait `T<t>` as a
concrete trait `T<int>` would not properly translate the runtime type
parameter `t` when used in the methods of the trait. Test
`traits_polymorphic.enc` has been updated to cover these cases.

The solution tries to be general enough so that only one place needs to
be changed when future AST nodes are added.