Fix initialization of type vars in closures (#660)
* Fix initialization of type vars in closures

This commit fixes a bug where type variables were not properly captured
in closures. The fix was the first change in `Closure.hs` and the change
in `encore.h` (allowing runtime types to be `NULL`). The rest of the
changes adds an assert that the runtime type is indeed initialized in
the beginning of running a closure. The generated code has been manually
checked.

Fixes #654.

* Move assert into runtime lib