replace "_init" and "init" by constructorName
I believe that checking for "init" was not a good idea since the
desugarer transforms every "init" method name into an "_init" method
name. for this reason, i have removed "init" in phases that work after
the desugarer. i have removed the hardcoded "init" and "_init"
appearances by a function `constructorName` that contains this value