refactor method calls and fix init msg send (#665)
* refactor method calls and fix init msg send

this commit refactors and unifies the notion of message sends and method
calls under the same function. it also forbids the typchecking of
sending an asynchronous self call the to the method `init` in the `Main`
class, i.e.

    class Main
      def main(): void
        this!main()    -- does not typecheck, which is good!

* fix suggestions from elias

* replace "_init" and "init" by constructorName

I believe that checking for "init" was not a good idea since the
desugarer transforms every "init" method name into an "_init" method
name. for this reason, i have removed "init" in phases that work after
the desugarer. i have removed the hardcoded "init" and "_init"
appearances by a function `constructorName` that contains this value

* Call constructor "init" internally