Support for type inference in constructor calls
This commit adds support for inference of type parameters when creating
new instances of classes. As long as all type parameters are used by the
constructor of the corresponding class, they should be inferable:

```
class Foo<t, u>
  def init(x : t, y : u) : void
    ()

class Main
  def main() : void
    new Foo(42, "Foo")
```

Before this commit, the last line would have had to be written as
´new Foo<int, String>(42, "Foo")´.

The implementation piggy-backs on the recently added support for
inference of type parameters to polymorphic methods.