add reduce, distinct and intersection combinators (#679)
* add reduce, distinct and intersection combinators

this commit improves the reduce combinator (which was blocking all
futures in the ParT) by reducing the number of blocking
operations (happen only for Future ( Par t)). it also adds two new
combinators, intersection and distinct, which performs the intersection
on the items in the ParT and the extraction of distinct items from the
ParT. To do this, it relies on the creation of a promise and a new
algorithm that waits until all futures in a ParT have been fulfilled,
using a single `await` statement and a lock-free implementation (instead
of `awaiting` in each future in the ParT).

NOTES: the reduce operation has two variants, one that considers that
the reduce function is associative, if the types are the same:

    reduce :: Par t -> t -> (t -> t -> t) -> Par t

and another one that cannot make this assumption and needs to run
sequentially:

    reduce :: Par t -> a -> (t -> a -> a) -> Par a

it is an implementation detail that the associative reduction is
implemented in terms of the sequential one. as soon as tasks are back
into Encore I will update this and other combinators.

* add derived combinator unionSet

* follow albert advice on

* add test case for union combinator

* rename callback function

* removed external dependency