refactor: ParT is a sequence
some combinators should guarantee an order, so that new combinators can
rely on that order. for instance, the sequence combinator:

    >> :: Par t -> (t -> t') -> Par t'

may perform operations in parallel but the end result should be expected
from the input. The following shows an example:

    val p = liftv(1) || liftv(2);
    extract(p >> \(x: int) -> x + 1)

expects results `[1, 2]` instead of `[2, 1]`. By maintaining the order
of some operators, the developer can reason about other combinators such
as `zip :: Par t -> Par t' -> (t -> t' -> t'') -> Par t''`.

this commit also improves the traversal of the set (in depth) by not
adding empty values to an auxiliary list used for the
traversal (implementation detail).