allow reentrant (recursive) mutex calls
this is needed by the join combinator. the case is the following:

**scenario**:
a method call that returns `Fut (Par t)`, lifted to `Par (Part)` and is
then _joined_ (`Par t`).

the runtime function `party_extract` creates a closure that puts in its env
the `Fut (Par t)` and awaits until all futures in the ParT are
fulfilled, in this case one future, and attaches a closure to run when
the futures are fulfilled. it does this by using future chaining. in the
case of chaining before the future (`Fut (Par t)`) is fulfilled, the
following happens:

  - fulfilled of the future grabs its lock on the future.
  - fulfils the future and checks on its chained closures and runs them.
    this closure contains a reference to the original `Fut (Par T)` and
    as soon as it tries to get the value from the future (which we
    already no that it's been fulfilled), it will try to grab the lock
    again to read that the future is in fact fulfil. this causes a
    reentrant call on the same lock that was acquired in the first point.