ParT contains counter to number elements
the ParT structure contains now a counter with the number of
elements. this counter is not precise as there may be `Fut (Par t)`
computations that are still on the go. if the ParT contains any of these
computations then the number of items needs to be counted and the
operation will block until all the items in the `Fut (Par t)` are fulfilled.