For comprehensions and iterators
Includes:
 - For comprehensions
 - documentation for the encore book (in the comments below)
 - iterator trait and implementations
 - adding methods to arrays, maybes, and futures (extensible!), without boxing [1]

[1] :: This commit adds methods to arrays and futures, without wrapping them
in extra boxes. This commit touches the type checker, the optimiser, and
adds a new top level module (called `BuiltinExt`) that contains the
implementations of array-methods.

`BuiltinExt` contains a bunch of implementation functions whose name starts,
e.g., with `array_`. A function called `array_foo` defines the method `foo` of
arrays. Each of these implementation functions gets one extra argument, of array
type, as its first argument. Additionally, there is a *trait* Array with similar
methods (but no extra arguments), that the type checker will use to typecheck
calls against arrays. Check the module docs for more details.

In the optimiser, we add a pass that switches all method calls on arrays,
futures, maybes (which did not extist before this commit) to function calls of
the respective implementation functions in `BuiltinExt`.