This PR adds the Eq and Id traits to Encore.
The Eq and Id standard traits are found in Std.enc which is included
by default, just like String, by all Encore programs.

Id comparison is supported on a type t iff
- t is a primitive type
- t is Maybe[t'] and t' supports id comparison
- t is a tuple (t1,...,tn) and t1..tn all support id comparison
- t is a reference type which includes the Id trait
- t is a type variable (sort-of unsound, see below!)

For now, the Id trait is empty and simply serves as a marker for
allowing the == operation. In the future, this may change.

Because we do not yet support type constraints, we have two
options for equality operations on type variables:
1) Disallow it for all type variables
2) Allow it for all type variables

This PR chooses 2) to not have to deprecate any existing tests. In
the future, once type constraints are added to type variables,
tests that rely on identity comparison on type variables must be
updated.

Also note that we don't support id comparisons across values of
different types modulo subtyping. Thus, if `a:uint` and `b:int`,
`a == b` is rejected by the compiler (etc).

The Eq trait is parameterised over a type t and requires a method
`def eq[t](other:t) : bool`. An example use is this:
```
class Cell[t] : Eq[Cell[t]]
  val value : t
    def eq(other:Cell[t]) : bool
        this.value == other.value
	  end
	  end
	  ```
	  Note that it relies on the unsound comparison on type
	  variables.

As a consequence of this change, some tests have been refactored
to only include valid id comparisons. Some standard library
components have been altered to include Id (and Eq) and/or to
make comparisons in different ways.

Id comparison is implemented via desugaring through matching. As a
consequence, id comparison may cause allocation as a side-effect
until tuples and maybes are allocated on the stack.