Parallelise tests.
On my machine, the time it takes to run `make test` goes down from 86sec to
46sec (speedup: x1.9).

This runs the test suites in parallel (granularity: per test suite).
Parallelising per test might lead to better load balancing (as test suites are
unevenly sized), but would require more invasive changes. A better way to
improve performance further would be to split up the `basic` test suite into
several smaller ones, I think.

The commit also splits basic test suite into smaller, topical suites to allow
for better parallelisation granularity (and because the basic suite is a dumping
ground anyway).

The bin/test script got the --do_seq flag for forcing a sequential run. When the
flag is not provided, the script will run in parallel mode. In parallel mode, if
GNU parallel is installed, the test suites will be executed in parallel with
each other. In parallel mode, if GNU parallel is not installed, the script will
print a warning and fall back to sequential mode.