Remove support for matching methods (#708)
This commit removes the remnants of matching functions and
methods from the compiler, i.e. having

```
fun fact(0) : int
  1
  | fact(n : int) : int
  n * fact(n - 1)
end
```
desugar into
```
fun fact(n : int) : int
  match n with
    case 0 => 1
    case n => n * fact(n - 1)
  end
end
```

This has not been implemented with the new syntax, and while it is
a neat feature, I think that we will be doing ourselves a favour
by removing this complexity from the compiler.