This PR adds a continue keyword to Encore
This PR adds a continue keyword to Encore which is usable inside

- while
- do/while
- repeat

but not for. The reason why they are not supported inside for
loops is their upcoming deprecation when for loops are replaced
by iterators.

Four tests added. Changed desugaring of repeat loops to get them
to work properly.

Addressed Elias' comments