remove desugaring of ParT function calls (#716)
* remove desugaring of ParT function calls

the ParT combinators `each`, `join`, `liftv`, `liftf` and `extract` are
not written as keywords in the parser but are being treated in a special
fashion at the desugarer level. this commit removes this desugaring and
adds these combinators as normal functions. the combinators have been
placed in the ParT module.

* add TODO on AST nodes that can be removed

* remove explicit types