Support for bounded mode polymorphism
This commit adds support for adding bounds on the mode of a type
parameter. It also introduces the mode `sharable` which abstracts
over all modes that may be safely shared between active
objects (`read`, `active` and `shared`). This mode can only be
used for type parameters for now, where it allows a polymorphic
context to pass values between active objects.