Bounded mode polymorphism (#776)
* Support for bounded mode polymorphism

This commit adds support for adding bounds on the mode of a type
parameter. It also introduces the mode `sharable` which abstracts
over all modes that may be safely shared between active
objects (`read`, `active` and `shared`). This mode can only be
used for type parameters for now, where it allows a polymorphic
context to pass values between active objects.

* Refactoring

* Reacted to @supercooldave's comments

* Only allow modes for type parameters

Changed parsing so that type variables can only have modes in the
parameter list of functions/classes/etc.

* Remove old test