Support for borrow blocks (#772)
* Support for borrow blocks

This commit adds support for borrow blocks:

```
borrow e as x in -- e : t
  ... -- x : borrowed t
end
```

When `e` is linear, the whole path must consist only of linear
targets (same rule as for `borrowed` function parameters), and the root
of the path may not be accessed inside the block.

The desugaring of `?.` and `?!` has been changed to first borrow the
target, so that it may be used on linear targets as well. This was the
main motivation for this commit.

* Disallow re-borrowing from borrowed variables

See the test `doubleBorrow.enc` for an example that would otherwise be
unsound.