Disallow silently capturing an active this as a passive reference
Before this commit, in an `active` class, you could use `this` to
access fields or (synchronously) call methods inside a closure.
Because the mode of `this` is `active`, which is safe, the mode of
the closure would allow it to be shared, leading to potential
data-races. This commit changes an `active` `this` to a `local`
variable if it is used for synchronous accesses. This forces the
`local` mode on a closure containing e.g. `this.foo()`.

There are also some improved error messages and a small check for
when a required `var` field is provided as a `val` field.

Fixes #781.