Add range information to parsed nodes
This commit extends the source position to include the range of
positions of an Encore AST-node. Before, when parsing an expression,
e.g. `(2,3)`, we would only know where the expression began. Now we also
have information on where the expression ends (in this example 5
characters after the beginning).

This is not visible to the programmer in any way, but is integral for
the upcoming editor mode support.