Add implicit constructor call in main class
This commit adds an implicit call to `init` at the beginning of
the `main` method. It also adds a restriction from defining a
constructor method in the `Main` class. "Why this madness?" I hear
you ask. This is to give field initialization a place to be
desugared into, while still keeping `main` as the starting point
of the program.