Allow id comparison of abstract traits
The only time an abstract trait appears is as the type of `this`
when typechecking an overridden method. This commit allows
comparison using `==` on *all* abstract trait types, with the
motivation that if a disallowed comparison was made it would be
caught when typechecking the class that includes the trait.