add task pool for loop combinator
add a Task pool for each use of the loop combinator. the task pool uses
a round-robin lock-free algorithm to dispatch work to each worker. it
loads balances a task to each actor. this may not be the best strategy
if an actor gets more complex tasks. if necessary, we can fix this in
the next iteration. load balancing with relaxed memory semantics.