prune streams its first value (lock-free alg.)
prune can be seen as a ParT that streams a single value.

    prune :: (Fut[Maybe[t]] -> Par[t']) -> Par[t] -> Par[t']

given its signature, we can understand that the ParT given as second
argument streams the first value out of it to the future in the first
argument. after that, we proceed by trying to kill the remaining
computations. the killing part is not finished yet.