Have active classes automatically include the Actor trait
The trait `Actor` is defined in the `Std` module, and defines
three methods: `await`, `suspend`, and `run_attached_closure`.
Automatically including this in active classes replaces the old
behaviour of implicitly defining the two former methods in active
classes.

The changelog looks big for such a change. This is because I also
fixed a bug that would incorrectly assume that a modeless,
trait-less, class was an active class (and thus included the
`Actor` trait). This means that the `Main` class now _has to_ be
declared as `active`, which some test cases did not.