Add loop and initial design of prune combinator (#857)
* add loop combinator to ParT

* add task pool for loop combinator

add a Task pool for each use of the loop combinator. the task pool uses
a round-robin lock-free algorithm to dispatch work to each worker. it
loads balances a task to each actor. this may not be the best strategy
if an actor gets more complex tasks. if necessary, we can fix this in
the next iteration. load balancing with relaxed memory semantics.

* prune streams its first value (lock-free alg.)

prune can be seen as a ParT that streams a single value.

    prune :: (Fut[Maybe[t]] -> Par[t']) -> Par[t] -> Par[t']

given its signature, we can understand that the ParT given as second
argument streams the first value out of it to the future in the first
argument. after that, we proceed by trying to kill the remaining
computations. the killing part is not finished yet.

* add TODO

* comment out testing code

* add tests to prune and loop